+++ src/main/java/kr/codesqaud/cafe/article/repository/ArticleRepository.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import kr.codesqaud.cafe.article.domain.Article;
-import kr.codesqaud.cafe.article.domain.Reply;
 
 public interface ArticleRepository {
 
@@ -17,12 +16,4 @@ public interface ArticleRepository {
 	void updateArticle(Article article);
 
 	void deleteArticle(Long idx);
-
-	void saveReply(Reply reply);
-
-	List<Reply> findAllReply(Long idx);
-
-	void deleteReply(String id, Long replyIdx);
-
-	String findReplyIdByIdx(Long replyIdx);
 }
