+++ src/main/java/kr/codesqaud/cafe/global/mapper/ArticleMapper.java
@@ -3,13 +3,10 @@ package kr.codesqaud.cafe.global.mapper;
 import org.springframework.stereotype.Component;
 
 import kr.codesqaud.cafe.article.domain.Article;
-import kr.codesqaud.cafe.article.domain.Reply;
 import kr.codesqaud.cafe.article.dto.ArticlePostRequest;
 import kr.codesqaud.cafe.article.dto.ArticleResponse;
 import kr.codesqaud.cafe.article.dto.ArticleTitleAndContentResponse;
 import kr.codesqaud.cafe.article.dto.ArticleUpdateRequest;
-import kr.codesqaud.cafe.article.dto.ReplyRequest;
-import kr.codesqaud.cafe.article.dto.ReplyResponse;
 
 @Component
 public class ArticleMapper {
@@ -33,13 +30,4 @@ public class ArticleMapper {
 		return new ArticleTitleAndContentResponse(article.getTitle(), article.getContent());
 	}
 
-	public Reply toReply(ReplyRequest replyRequest) {
-		return new Reply(replyRequest.getId(), replyRequest.getArticleIdx(), replyRequest.getNickName(),
-			replyRequest.getContent());
-	}
-
-	public ReplyResponse toReplyResponse(Reply reply) {
-		return new ReplyResponse(reply.getNickName(), reply.getContent(), reply.getDate(), reply.getArticleIdx(),
-			reply.getReplyIdx());
-	}
 }
