+++ src/main/java/kr/codesqaud/cafe/reply/domain/Reply.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.article.domain;
+package kr.codesqaud.cafe.reply.domain;
 
 import static kr.codesqaud.cafe.global.util.DateUtil.*;
