+++ src/main/java/kr/codesqaud/cafe/reply/dto/ReplyRequest.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.article.dto;
+package kr.codesqaud.cafe.reply.dto;
 
 import javax.validation.constraints.NotBlank;
