+++ src/main/java/kr/codesqaud/cafe/reply/dto/ReplyResponse.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.article.dto;
+package kr.codesqaud.cafe.reply.dto;
 
 public class ReplyResponse {
 	private final String nickName;
